Reject invalid wildcards in alternatives.
This change detects and rejects invalid tokens in alternative branches.
In particular, tree and zero-or-more tokens in alternative branches may
be invalid if they are adjacent to certain tokens surrounding their
associated alternative. For example, the following patterns were
previously accepted but are now rejected: `{**}`,
`prefix{**/error,okay}`, `{**/okay,error/**}postfix`, and
`{okay,error$}*`.

A `rule` module has been introduced for checking a sequence of tokens
for correctness and extension traits for slices and iterators have been
added to (or moved into) the `glob` module.