Use `--quit-on-intr` by default.
This change includes the `--quit-on-intr` flag in the default `less`
pager command line. This kills both the pager child process and parent
process when an interupt signal is received by `less`.

Without this option, it is easy for a defunct (zombie) `less` process to
remain after using `^C` to interupt a paging `nym` command on Linux.