Partition globs in from-patterns.
This change uses partitioned globs in from-patterns. This allows the
platform to re-interpret literal path prefixes, such as parent
directories `..`. For example, `../**/*.rs` will read from the parent of
the current working directory. Note that `..` is not special; it is
parsed as a literal. This only works for literal prefixes, so a glob
pattern like `**/../*.rs` has very little utility and will only match
paths that contain the literal component `..`.

This change also fixes a bug in partitioning wherein tree tokens were
incorrectly indexed.